From http://www.jwz.org/xscreensaver/xscreensaver-5.31.tar.gz
-rw-r--r-- 1 zblaxell zblaxell 10569562 Nov 15 18:51 xscreensaver-5.31.tar.gz
1b023a3b01f1f23aa136d13c37f79c0c01c3f1e6  xscreensaver-5.31.tar.gz

+++ hacks/glx/stonerview-move.h
@@ -14,9 +14,9 @@
 #define __STONERVIEW_MOVE_H__
 
 typedef struct {
-  GLfloat pos[3];
-  GLfloat vervec[2];
-  GLfloat col[4];
+  double pos[3];
+  double vervec[2];
+  double col[4];
 } stonerview_elem_t;
 
 extern void stonerview_init_move(stonerview_state *);
